+++ R/admixcor.R
@@ -8,6 +8,7 @@ admixcor <- function(
                      L_type = c('identity', 'uniform', 'diagrandom', 'random'),
                      Q_algorithm = c('original', 'nnls', 'bvls', 'glmnet', 'quadprog', 'quadprog-compact'),
                      L_algorithm = c('original', 'nnls', 'bvls', 'glmnet'),
+                     kronecker = FALSE,
                      tol = sqrt( .Machine$double.eps ), # 1e-15
                      nstep_max = 100000,
                      report_freq = 1000
@@ -61,7 +62,7 @@ admixcor <- function(
         
         # apply the updates, one at the time
         R1 <- update_R( ThetaSR, Q0, L0 )
-        L1 <- update_L( ThetaSR, Q0, R1, gamma, I, algorithm = L_algorithm )
+        L1 <- update_L( ThetaSR, Q0, R1, gamma, I, algorithm = L_algorithm, kronecker = kronecker )
         Q1 <- update_Q( ThetaSR, L1, R1, delta, I, algorithm = Q_algorithm )
         
         # calculate step sizes, to assess convergence